# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample Pet Store App",
        "description": "This is a sample server for a pet store.",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1",
            "description": "Localhost"
        }
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "schemas": {
            "id": {
                "components": {
                    "id": {
                        "type": "string"
                    }
                }
            },
            "token": {
                "components": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "users": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userName": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "gmail": {
                        "type": "string"
                    },
                    "linkAvatar": {
                        "type": "string"
                    },
                    "statusActive": {
                        "type": "string"
                    },
                    "point": {
                        "type": "number"
                    },
                    "port": {
                        "type": "number"
                    },
                    "numberOfWins": {
                        "type": "number"
                    },
                    "numberOfLosses": {
                        "type": "number"
                    },
                    "resetPasswordToken": {
                        "type": "string"
                    },
                    "resetPasswordTokenExp": {
                        "type": "string"
                    }
                }
            },
...
            "listFriends": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "listID": {
                        "type": "array",
                        "items": {
                            "schema": {
                                "$ref": "#/components/schemas/users"
                            }
                        }
                    },
                    "status": {
                        "type": "string",
                        "default": "Waiting"
                    }
                }
            },
...
            "matches": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "players": {
                        "type": "array",
                        "items": {
                            "schema": {
                                "$ref": "#/components/schemas/users"
                            }
                        }
                    },
                    "status": {
                        "type": "string"
                    },
                    "count": {
                        "type": "number"
                    },
                    "betPoints": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/listFriends/edit/{id}": {
            "put": {
                "tags": [
                    "listFriends"
                ],
                "summary": "Cập nhật trạng thái là bạn bè theo id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of list friends",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/listFriends"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/listFriends/add": {
            "post": {
                "tags": [
                    "listFriends"
                ],
                "summary": "Khởi tạo danh sách bạn bè",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addFriend"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/listFriends"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/listFriends/delete/{id}": {
            "delete": {
                "tags": [
                    "listFriends"
                ],
                "summary": "Xóa bạn bè theo id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/listFriends"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/matches/add": {
            "post": {
                "tags": [
                    "matches"
                ],
                "summary": "Tạo phòng đấu",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addMatches"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/matches"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/matches/edit/addOrSubUser/{id}": {
            "put": {
                "tags": [
                    "matches"
                ],
                "summary": "Cập nhật phòng đấu",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "requestBody": {
                    "description": "option có 2 trường hợp là adduser (tham gia trận đấu), subuser (thoát khỏi phòng đấu)",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addOrSubUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/users"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/matches/edit/resultMatch/{id}": {
            "put": {
                "tags": [
                    "matches"
                ],
                "summary": "Cập nhật phòng đấu",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/resultMatch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/users"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/matches/delete/{id}": {
            "delete": {
                "tags": [
                    "matches"
                ],
                "summary": "Xóa phòng đấu",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of matches",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/matches"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Đăng ký tài khoản",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/register"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Đăng nhập tài khoản",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/login"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/forgotPassword": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Quên mật khẩu",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/forgotPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/getToken/{token}": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Nhập mã xác thực",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "input Token",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/resetPassword": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Đổi mật khẩu",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/resetPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/users"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}